pyproject.toml
[project]
name = "optimizer-comparison-cnn"
version = "0.1.0"
description = "Compare SGD, Adam and PSD optimizers on small CNN benchmarks"
readme = "README.md"
authors = [{name = "Optimizer Comparison Contributors"}]
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "torch",
    "torchvision",
    "matplotlib",
    "numpy",
    "psd-optimizer>=0.1.2",
    "tqdm",
    "pandas",
    "typer",
    "pyyaml",
]
[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "black",
    "isort",
    "mypy",
    "mkdocs-material",
    "pre-commit",
]

[project.scripts]
optimizer-cmp = "optimizer_comparison.cli:app"

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.mypy]
ignore_missing_imports = true
